Create separate test files for next & previous btn
