+++ pages/organisms/carousel/carousel.test.tsx
@@ -3,7 +3,6 @@ import "@testing-library/jest-dom";
 import { CarouselButton, CarouselData, CarouselItem } from "../../../data";
 import Carousel from "./carousel";
 import CarouselPreviousButton from "../../atoms/carousel-previous-button/carousel-previous-button";
-import CarouselNextButton from "../../atoms/carousel-next-button/carousel-next-button";
 
 const carouselItem: CarouselItem = {
   title: "Made with love",
@@ -49,38 +48,4 @@ it("Let's check carousel title, sub-title and image altTag", () => {
     expect(title).toBeInTheDocument();
     expect(imageAltTag).toBeInTheDocument();
   });
-});
-
-//better to test buttons from parent compoent then create seperate test file in child component
-
-it("Test 'next' button", () => {
-  const { getByRole } = render(<CarouselNextButton />);
-  const button = screen.getByRole("button", { name: "next button" });
-  expect(button).toBeInTheDocument();
-});
-
-it("Test 'previous' button", () => {
-  const { getByRole } = render(<CarouselPreviousButton />);
-  const button = screen.getByRole("button", { name: "previous button" });
-  expect(button).toBeInTheDocument();
-});
-
-it("Is 'previous' button clicked?", () => {
-  const clickPrevious = jest.fn();
-  const { getByRole } = render(
-    <CarouselPreviousButton loadPreviousSlide={clickPrevious} />
-  );
-  const button = screen.getByRole("button", { name: "previous button" });
-  fireEvent.click(button);
-  expect(clickPrevious).toHaveBeenCalledTimes(1);
-});
-
-it("Is 'next' button clicked?", () => {
-  const clickNext = jest.fn();
-  const { getByRole } = render(
-    <CarouselNextButton loadNextSlide={clickNext} />
-  );
-  const button = screen.getByRole("button", { name: "next button" });
-  fireEvent.click(button);
-  expect(clickNext).toHaveBeenCalledTimes(1);
 });
